requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-index-url requirements.in
#
attrs==22.1.0
    # via
    #   pytest
    #   pytype
autoflake==1.4
    # via -r requirements.in
black==22.6.0
    # via -r requirements.in
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via black
flake8==5.0.4
    # via -r requirements.in
google-api-core==2.8.2
    # via google-api-python-client
google-api-python-client==2.58.0
    # via -r requirements.in
google-auth==2.11.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==0.5.2
    # via -r requirements.in
googleapis-common-protos==1.56.4
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
importlab==0.7
    # via pytype
iniconfig==1.1.1
    # via pytest
libcst==0.4.7
    # via pytype
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via
    #   black
    #   typing-inspect
networkx==2.8.3
    # via
    #   importlab
    #   pytype
ninja==1.10.2.3
    # via pytype
oauthlib==3.2.0
    # via requests-oauthlib
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
protobuf==4.21.5
    # via
    #   google-api-core
    #   googleapis-common-protos
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycodestyle==2.9.1
    # via flake8
pydantic==1.9.2
    # via -r requirements.in
pydot==1.4.2
    # via pytype
pyflakes==2.5.0
    # via
    #   autoflake
    #   flake8
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
    #   pydot
pytest==7.1.2
    # via -r requirements.in
pytype==2022.8.23
    # via -r requirements.in
pyyaml==6.0
    # via libcst
requests==2.28.1
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
tabulate==0.8.10
    # via pytype
toml==0.10.2
    # via pytype
tomli==2.0.1
    # via
    #   black
    #   pytest
typing-extensions==4.3.0
    # via
    #   libcst
    #   pydantic
    #   pytype
    #   typing-inspect
typing-inspect==0.8.0
    # via libcst
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests
